requirements.txt
line-bot-sdk
google-auth
google-cloud-datastore
google-api-python-client
google-auth-oauthlib
google-cloud-vision